pyproject.toml
[project]
name = "mini-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "myapplication>=0.1.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "tomlkit>=0.13.3",
    "tortoise-orm[asyncpg]>=0.25.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "aerich>=0.9.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "ruff>=0.12.12",
    "tomli-w>=1.2.0",
    "tortoise-orm>=0.25.1",
]
prod = [
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]

[tool.aerich]
tortoise_orm = "app.db.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
